Refactored some code in Unity project
Former-commit-id: a612692681c347a9b7d5a1edb260370d477b17f2
Former-commit-id: 1f85098aa6f2e3276eed6164bc1263e2e899ad9e [formerly eab92eda1163224a880fb19dd0a44396b2a7e085] [formerly 2f5554147f29509845ff48b07d6bc4381928bfa6 [formerly 7b302e424ce224e0b95c655df77c6bd9a485e109]] [formerly 5fce0311a303a53a066ac5f5bf121349ed74f71b [formerly 06f83cbb4d886a989a8b41b7745088757dfb7c69] [formerly 09650960b7495317552f86b848d35485587cdd8b [formerly 8c21241ce8e8d0597f7e30f4fa37a5061ad92e32]]] [formerly a0b601e1fef3741e5999e4144479d7c9a07df5a6 [formerly df97beef85448598231b70fd597c5e7b1fb5faeb] [formerly ca7df5cf601606734ac8df6a001c5afcad3614f2 [formerly 904b9c10c9190719d422bc6d82b861ffc83823ff]] [formerly 47da13979b2a916a528224f6f37c68fbf693735c [formerly 808675d5419a44fe4f9245b1736a6f51bd60ba64] [formerly 975d2cec803378853228c5843cd862e1b83fb377 [formerly 5569beda646b6df398faab6cf4358470a218020c]]]] [formerly 08e9894ca8037dd344a11a00e52ae427b1b88aa0 [formerly 2901e7d7baa4202fb1ac27094c315d30a0f621fd] [formerly 54c1de10efcacf66e354093e4941f67397ab61f6 [formerly 13376b99e12d3523b8a28121bacff8ab92a9c31d]] [formerly 6b2fe348129f92e68c099e895734ddc1d5468216 [formerly 71b22068798bc0e8b86e3d869c11914ee1b49416] [formerly a8587008e99a078fe5cb3ae722868a089962bae7 [formerly b46abc77ba3b62645ce307cfa7aa966754b3928b]]] [formerly 1f64aee5d5cad088e9541654ce2c9d4224a2cad4 [formerly 5fef516a24d52c3a70d07a0988f22883dec5f842] [formerly d7c413bdd1036f1d478c8b9d2264d0d4ff802398 [formerly 18766be0dab1d23467e5b0494db1f56f694881bb]] [formerly 05d8220e458350e2cff1c3f19c5881c9952867c4 [formerly 10caefff542f082265e590bd40d6fd37b997fe19] [formerly c3cc1dab42800a8fc0b85c51c971056e859ee80e [formerly 563a1612cc8f01b34bfb7ffd55f0efb1e5c91cc7]]]]] [formerly be368ee563a4c303427668b19e9e2c9d89cf7b30 [formerly 75ab376443c7bbdde8cb2d35cb0d4bd24dcc27cc] [formerly d1f1769f6c07bf2f3b909c035690dec3f891dae2 [formerly 0b294c56e78d0309438fa98b1c36cc481b96d1bf]] [formerly 439fda7d8b36df15a0038aa8bc94a3f8725c78ff [formerly d9fe7d854df5d1374265639623c001bcf11acdc2] [formerly 57c4f0088fd4dfb06a3383f784de38e571068b1b [formerly d39ee551c9b6acceec2c8ee46ac4c5f7cb79a4d5]]] [formerly df6e963eeb10cd5722949e64d10245fe64405658 [formerly a7814287755fa019d1a3e080455d6cccdeaf41d3] [formerly 753e69d73b324548bef99df8f3f639176979b89e [formerly 490edd67ef7302b7fb571294fa7098a6ebddfb4a]] [formerly 2771d0a61a0b461ed3fe0ccb04deeecf6b74ae62 [formerly 83d88e3c38b4b9cc576de95941d7addcf5dc32d8] [formerly bff29d79ab242b157f7e24a1192475023e753a21 [formerly 87ed690833f2c18c9fc840b7fcbc6a83cf916bb9]]]] [formerly 8b1ac4c5073b48d31f365fcdfcf40530ecfdc5b8 [formerly e2885fb9292b922e23de43b58dd47422ef64fe47] [formerly eee76b3289bbacf23cdd4adee8b3febc47b69e97 [formerly cf15316f882226583bdd741964487b2c1310d10e]] [formerly 11ed04143c2a136d073924abe806d72f85289901 [formerly e5b296f6b4650bda146a657a5125b263426e98bb] [formerly 61d43c4c6e2f2c1a5865d1a5f192dd41de23104b [formerly 2fd09652756db18394723955f7a207b2b54742fa]]] [formerly 6423cb47b875d62d14eb88e7d185ae2e304bd62c [formerly 26e95cc197c31117f8679d737a516db155cf552e] [formerly 2f58cd3b49085227ea23855b12b13e754b2a4329 [formerly 2d161404da5e355744bf832e42b7c54d94807de4]] [formerly 2c5dd1e1730152d4c453cdbb4cc0e8157234af74 [formerly bababbeb30e061c48eb034ac3206dcf5d007960c] [formerly d54b5475b034c332f6ab44ffc96b8ee3ea0ec9be [formerly 441fefae4e2403a412d0421ec9f136e121407b7d]]]]]]
Former-commit-id: b3b2890a5d368e1bb674c2f717694ec1cd71b700 [formerly 90407c9434a4b08e51dd83242f1c991c99b6072b] [formerly d69501bb56deecdd803994f2fc51ed4d6abdb1bf [formerly 38eb168c43830a2e19284e7eff518e143f38abeb]] [formerly 1e78744c35f21a56e393a5f076b5a86baf4604e8 [formerly 01c31f34ea4322d4d779a28b78b053f76f67c821] [formerly e77db2cc6e044fd33bdb956e6a0e1eb3eee2f19b [formerly 2457eedbbb0546326e2aae2997d2d02c4170e9ce]]] [formerly d835048abe171b2738d3e8df3a4c29c88460a1b9 [formerly 17a9197758377dc29130c5b81dfce7b671830d2e] [formerly 34011493688dbafd2f0c7a811932db59a2eb6d39 [formerly 828a356cf9b126b4e87202d2c6dd039b16f39cf0]] [formerly 954f1005b9ab95b0ce937f2474dbb43803fc241f [formerly 9ac72e0640fd4b3dfcf7ebedefa8b6c2f09bab27] [formerly 7bd02d66f95c4cc4d44babd23ed68872b2383630 [formerly d8e20df1c0bb225c9f2dd5f8a54969f2e1762d77]]]] [formerly e6f160132f3573b25e19f28e7f1ec41251d4d060 [formerly 17eebc45babbaf550ac150fd5ef725142fb67698] [formerly 0d9e2d9f2c121d92f5aba162317fc319bf891e9d [formerly 393d8508965d852697cefb9526aef3cf918acae7]] [formerly 246c2fcb4e91fd8fef2cef95e091268e35a7932b [formerly 2e51a4d84d01a76fb085928b266449ad1bf3a327] [formerly 28268d0fd7db4ad20acb2f91a83fbd0d5dd30f99 [formerly eae59d2455dc015f5df108a78c417b139c712845]]] [formerly 3aa4d77f2d793915e11ec54b3cf72f4f78499670 [formerly 5838c6a7f4e58cc711e44070960d325aa84ea952] [formerly 7047a321c31182be5bc180958aa4ca8389df0a39 [formerly 2a7f7829992a237591027aa259b2868d7f62d12b]] [formerly 95e3c1071dc52d191592607ff6ba5e1807e8d0f5 [formerly ee6703b3f5680fcfa9ff920da2d27c25ca5c658c] [formerly 3b37ff530acaf39d80c41078a38f01edb3776449 [formerly b6f61a5947d96cbef755a10d59b59a59dd43c6f3]]]]]
Former-commit-id: 4e38194fde3c35a92105ed95e19059adb6cb0569 [formerly 414ec19122e6c028623e31b6d1170949612d6c3b] [formerly 3be75ada0d334af6923fec91ef566bff4f3471cd [formerly 991bcc6a9885336a4f8859a57f8cb582d05090f4]] [formerly e3391dfef2f62de00cc51383257fb8dec83753ef [formerly 7b43ec68085d76aad26ddfcf936c44a7d05d3fcc] [formerly 41326318cec438d825cb073746fdbf62e400f12d [formerly 58a5ab423f52ff0d9225ae8f4b8ab2caa667393a]]] [formerly 9727db764244477005d31a3d80729ba92636188e [formerly 50fc1a19f99e6dc1a0b74042f261586c8e113eea] [formerly 9cbd0d280dee4e93e416c1b80f4849c0a5426707 [formerly d13173f36df559266d09dcf7d7dca18bbd29dc80]] [formerly d39861ef97e9d53dad7bacda4ef23147b5d646b1 [formerly fe15521baea051fa48c68d449b80c0945993133e] [formerly 9452572e0981770217c202d8ddf0f744cb48c106 [formerly 190a4b859fdc64b880334a96caba91dfb3e99980]]]]
Former-commit-id: 4721d20103ff3b416b9bd805ef63149b63270c34 [formerly 79add9ba4f2a6f93394d040c866674d7727294bc] [formerly 4703b08e0fe95cf8dc79df397fb448f2237c071e [formerly bda1495a04142a3a44cbd69cbcc4e2f52b6580ef]] [formerly 192044fa1fad36e84fc98ab2f63bd87f9f31e79d [formerly f8b9aff8712cf3528ff490b53e0fda3ec26058a8] [formerly 890f7e1f6fbbdb9831afa4d1536576aeb6649285 [formerly 7d8796bdb99293a0462c8c613860cdb9deb96106]]]
Former-commit-id: 7ee63c7838b8542eda65acbee2f673fff8b99473 [formerly 3d2d785781321c5090fed3bdd144057c6eb753c3] [formerly 36f1918918f0136ffbcf5440dd66282a53c4d980 [formerly 77ecbecd449c3288c14e41fc57212a0e54976196]]
Former-commit-id: 0669f9e363e9c0c164e417f16326cddf56e49751 [formerly 1ca4352cbd60e1689be31989ca67e12bd59b6468]
Former-commit-id: bbd686e50884c947519dce0459c2ba6cd78481b2

+++ Wheat Simulation/Assets/Scripts/Python/updatejson.py
@@ -16,14 +16,14 @@ def json_copy(input_json_path, output_json_path, only_wheat=False):
     with open(input_json_path, 'r') as f:
         data = json.load(f)
 
-    if only_wheat:
-        # Filter annotations
+    # Filter annotations
+    if not only_wheat:
+        filtered_annotations = [anno for anno in data['annotations']]
+    else:
         filtered_annotations = [anno for anno in data['annotations'] if anno['category_id'] == 1]
-        data['annotations'] = filtered_annotations
 
-        # Filter categories
-        filtered_categories = [cat for cat in data['categories'] if (cat['name'] == "Ground" or cat['name'] == "Head")]
-        data['categories'] = filtered_categories
+    # Update the data dictionary
+    data['annotations'] = filtered_annotations
 
     # Write the new JSON file
     with open(output_json_path, 'w') as f:
@@ -35,18 +35,18 @@ def reorganize_dataset(domain_path):
     if os.path.exists(domain_path):
         # Make dir for the release dataset
         domain_path = domain_path.rstrip('/\\')
-        new_domain_path = f"{domain_path}_seg"
-        if not os.path.exists(new_domain_path):
-            os.mkdir(new_domain_path)
+        new_directory_path = f"{domain_path}_seg"
+        if not os.path.exists(new_directory_path):
+            os.mkdir(new_directory_path)
 
         # Copy the JSON into the new directory
         json_copy(f"{domain_path}/coco_annotations.json",
-                  f"{new_domain_path}/coco_annotations.json")
+                  f"{new_directory_path}/coco_annotations.json")
         json_copy(f"{domain_path}/coco_annotations.json",
-                  f"{new_domain_path}/coco_annotations_onlywheatheads.json", True)
-        
+                  f"{new_directory_path}/coco_annotations_onlywheatheads.json", True)
+
         # Copy the images into the new directory images sub-folder
-        new_images_path = f"{new_domain_path}/images"
+        new_images_path = f"{new_directory_path}/images"
         os.mkdir(new_images_path)
         images = [image for image in glob.glob(f"{domain_path}/*_image.png")]
         for image in images:
@@ -58,7 +58,7 @@ def reorganize_dataset(domain_path):
 
 
 # Main code
-def process_data(domain_path):
+def get_polygon_segmentation_from_masks(domain_path):
     if os.path.exists(domain_path):
         # Make dir for the release dataset
         domain_path = domain_path.rstrip('/\\')
@@ -123,13 +123,81 @@ def process_data(domain_path):
             if annotation_id in segmentations:
                 annotation['segmentation'] = segmentations[annotation_id]
 
-        # Fix the bounding boxes and add "iscrowd=0"
-        coco_data = add_is_crowd_and_fix_bbox(coco_data)
+        # Test with standard JSON module
+        standard_json_file_path = os.path.join(domain_path, 'coco_annotations.json')
+        try:
+            with open(standard_json_file_path, 'w') as f:
+                json.dump(coco_data, f, indent=2)
+            print(f"Standard JSON serialized and saved to {standard_json_file_path}")
+        except Exception as e:
+            print(f"Error saving JSON file with standard json module: {e}")
 
-        # Clear invalid segmentations
-        coco_data = clear_invalid_segmentations(coco_data)
+def get_rle_segmentation_from_masks(domain_path):
+    if os.path.exists(domain_path):
+        # Make dir for the release dataset
+        domain_path = domain_path.rstrip('/\\')
+        # Load the existing JSON file
+        json_file_path = os.path.join(domain_path, 'annotations.json')
+        with open(json_file_path, 'r') as f:
+            coco_data = json.load(f)
 
-        # Save with standard JSON module
+        # Remove the "annotationMaps" section
+        annotation_maps = coco_data.pop('annotationMaps', [])
+
+        print(f"Total annotation maps: {len(annotation_maps)}")
+
+        # Dictionary to hold segmentation data
+        segmentations = defaultdict(list)
+
+        # Process each annotation map with a progress bar
+        for annotation_map in tqdm(annotation_maps, desc="Processing annotation maps"):
+            annotation_file = annotation_map['file_name']
+
+            # Construct the full path to the annotation PNG file
+            annotation_file_path = os.path.join(domain_path, annotation_file)
+
+            # Read the annotation PNG file, collapse the four channels to a 32-bit integer
+            annotation_img = cv2.imread(annotation_file_path, cv2.IMREAD_UNCHANGED)
+            if annotation_img.shape[2] == 4:
+                # Extract each channel
+                r_channel = annotation_img[:, :, 2]  # Red channel
+                g_channel = annotation_img[:, :, 1]  # Green channel
+                b_channel = annotation_img[:, :, 0]  # Blue channel
+                a_channel = annotation_img[:, :, 3]  # Alpha channel
+
+                # Combine channels into a single 32-bit integer
+                annotation_img = (r_channel.astype(np.uint32) << 24) | \
+                                 (g_channel.astype(np.uint32) << 16) | \
+                                 (b_channel.astype(np.uint32) << 8) | \
+                                 (a_channel.astype(np.uint32))
+
+            if annotation_img is None:
+                print(f"Failed to load annotation image: {annotation_file_path}")
+                continue
+
+            # Convert the annotation image to unique object IDs
+            unique_ids = np.unique(annotation_img)
+
+            for obj_id in unique_ids:
+                if obj_id == 0:
+                    continue  # Skip the background
+                # print(obj_id)
+
+                # Create a binary mask for the current object ID
+                binary_mask = annotation_img == obj_id
+
+                polygons = Mask(binary_mask).polygons() # add the new seg method here
+
+                for polygon in polygons:
+                    segmentations[obj_id].append(polygon.tolist())
+
+        # Update the annotations in the COCO data with the segmentation data
+        for annotation in tqdm(coco_data['annotations'], desc="Updating Annotations"):
+            annotation_id = annotation['id']
+            if annotation_id in segmentations:
+                annotation['segmentation'] = segmentations[annotation_id]
+
+        # Test with standard JSON module
         standard_json_file_path = os.path.join(domain_path, 'coco_annotations.json')
         try:
             with open(standard_json_file_path, 'w') as f:
@@ -139,51 +207,56 @@ def process_data(domain_path):
             print(f"Error saving JSON file with standard json module: {e}")
 
 
-def add_is_crowd_and_fix_bbox(data):
-    resolution = 1024
+# Copy a JSON file to a new location. Categories can be excluded from the final result by passing a list of ints.
+def add_is_crowd_and_fix_bbox(domain_path):
+    json_path = f"{domain_path}coco_annotations_onlywheatheads.json"
 
+    # Load the existing JSON file
+    with open(json_path, 'r') as f:
+        data = json.load(f)
+
+    # Remove unused categories
+    filtered_categories = [cat for cat in data['categories'] if (cat['name'] == "Ground" or cat['name'] == "Head")]
+    data['categories'] = filtered_categories
+
+    resolution = 1024
     # Fix the bounding box format and add iscrowd
     for annotation in data['annotations']:
-        # Calculate new bbox
         bounding_box = annotation['bbox']
-        bl_x = bounding_box[0] # bottom left x value
-        bl_y = bounding_box[1] # bottom left y value
+        bl_x = bounding_box[0]
+        bl_y = bounding_box[1]
         tr_x = bounding_box[2]
         tr_y = bounding_box[3]
 
-        width = tr_x - bl_x # save as [top left pixel x, top left pixel y, width, height]
+        width = tr_x - bl_x
         height = tr_y - bl_y
 
         annotation['bbox'] = [bl_x, resolution - tr_y, width, height]
         annotation['iscrowd'] = 0
 
-    return data
+    # Write the new JSON file
+    with open(json_path, 'w') as f:
+        json.dump(data, f, indent=2)
 
-def clear_invalid_segmentations(data):
-    # For each annotation, remove invalid segmentations (polygons with less than three vertices)
-    for annotation in data['annotations']:
-        annotation['segmentation'] = [segmentation for segmentation in annotation['segmentation'] if len(segmentation) >= 6]
-    
-    # Remove any annotation that has no valid segmentations
-    data['annotations'] = [annotation for annotation in data['annotations'] if len(annotation['segmentation']) > 0]
 
-    return data
-
-# Call the other functions using the argument given
-if sys.argv[1] is not None:
+if sys.argv[1] is not None and sys.argv[2] is not None:
     dataset_path = sys.argv[1]
     domain_paths = glob.glob(f"{dataset_path}/*/")
+    print(f"Domain paths found: {domain_paths}")
 
     if len(domain_paths) > 0:
-        print(f"Domain paths found: {domain_paths}")
         for domain_path in domain_paths:
             domain_path = domain_path.rstrip("/\\")
 
-            # Copy the original "annotations.json", process to add polygon segmentation, save as "coco_annotations.json" in the same directory
-            process_data(domain_path)
+            # polygon encoding segmentation
+            get_polygon_segmentation_from_masks(domain_path)
 
-            # Create a new directory with a different name, organized with images in one folder and two annotation files (one for all objects and one for only wheat)
+            # Move the files to another directory without anything unnecessary
             reorganize_dataset(domain_path)
 
+            # Fix some issues with the annotation
+            add_is_crowd_and_fix_bbox(domain_path)
     else:
-        print("Error: No domain paths found under dataset directory.")
+        print("Error: could not find any domain paths in dataset path")
+else:
+    print("arg 1: dataset path (not domain path!)")
